{"schema":"decodex/commit/1","summary":"Extract scroll live stream owner","authority":"manual"}
